requirements.txt
asgiref==3.11.1
Django==6.0.2
django-bootstrap5==26.2
django-debug-toolbar==6.2.0
django-simple-history==3.11.0
sqlparse==0.5.5
tzdata==2025.3
